candidate management api (Updated)

+++ src/main/java/ma/nttdata/externals/module/candidate/service/impl/CandidateSrvImpl.java
@@ -1,5 +1,6 @@
 package ma.nttdata.externals.module.candidate.service.impl;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import ma.nttdata.externals.module.candidate.dto.CandidateDTO;
 import ma.nttdata.externals.module.candidate.entity.Candidate;
@@ -52,73 +53,16 @@ public class CandidateSrvImpl implements CandidateSrv {
     public CandidateDTO update(UUID id, CandidateDTO candidateDTO) {
         Optional<Candidate> existingCandidateOpt = candidateRepository.findById(id);
         if (existingCandidateOpt.isEmpty()) {
-            return null;
+            throw new RuntimeException("Candidate not found with id :" + id);
         }
         Candidate existingCandidate = existingCandidateOpt.get();
         Candidate updatedCandidate = mapper.candidateDTOToCandidate(candidateDTO);
-
         existingCandidate.setFullName(updatedCandidate.getFullName());
         existingCandidate.setBirthDate(updatedCandidate.getBirthDate());
         existingCandidate.setYearsOfExperience(updatedCandidate.getYearsOfExperience());
         existingCandidate.setGender(updatedCandidate.getGender());
         existingCandidate.setMainTech(updatedCandidate.getMainTech());
         existingCandidate.setSummary(updatedCandidate.getSummary());
-
-        Address updatedAddress = updatedCandidate.getAddress();
-        if (updatedAddress != null) {
-            Address existingAddress = existingCandidate.getAddress();
-            if (existingAddress == null) {
-                existingAddress = new Address();
-                existingCandidate.setAddress(existingAddress);
-            }
-            existingAddress.setStreet(updatedAddress.getStreet());
-            Country updatedCountry = updatedAddress.getCountry();
-            if (updatedCountry != null) {
-                Optional<Country> existingCountry = countryRepository.findByEnglishName(updatedCountry.getEnglishName());
-                if (existingCountry.isPresent()) {
-                    existingAddress.setCountry(existingCountry.get());
-                } else {
-                    Country savedCountry = countryRepository.save(updatedCountry);
-                    existingAddress.setCountry(savedCountry);
-                }
-            }
-            City updatedCity = updatedAddress.getCity();
-            if (updatedCity != null) {
-                Optional<City> existingCity = cityRepository.findByName(updatedCity.getName());
-                if (existingCity.isPresent()) {
-                    existingAddress.setCity(existingCity.get());
-                } else {
-                    City savedCity = cityRepository.save(updatedCity);
-                    existingAddress.setCity(savedCity);
-                }
-            }
-        }
-        if (updatedCandidate.getContacts() != null) {
-            existingCandidate.getContacts().clear();
-            existingCandidate.getContacts().addAll(updatedCandidate.getContacts());
-            updatedCandidate.getContacts().forEach(contact -> contact.setCandidate(existingCandidate));
-        }
-        if (updatedCandidate.getExperiences() != null) {
-            existingCandidate.getExperiences().clear();
-            existingCandidate.getExperiences().addAll(updatedCandidate.getExperiences());
-            updatedCandidate.getExperiences().forEach(exp -> exp.setCandidate(existingCandidate));
-        }
-        if (updatedCandidate.getSkills() != null) {
-            existingCandidate.getSkills().clear();
-            existingCandidate.getSkills().addAll(updatedCandidate.getSkills());
-            updatedCandidate.getSkills().forEach(skill -> skill.setCandidate(existingCandidate));
-        }
-        if (updatedCandidate.getEducations() != null) {
-            existingCandidate.getEducations().clear();
-            existingCandidate.getEducations().addAll(updatedCandidate.getEducations());
-            updatedCandidate.getEducations().forEach(edu -> edu.setCandidate(existingCandidate));
-        }
-        if (updatedCandidate.getLanguages() != null) {
-            existingCandidate.getLanguages().clear();
-            existingCandidate.getLanguages().addAll(updatedCandidate.getLanguages());
-            updatedCandidate.getLanguages().forEach(lang -> lang.setCandidate(existingCandidate));
-        }
-
         Candidate savedCandidate = candidateRepository.save(existingCandidate);
         return mapper.candidateToCandidateDTO(savedCandidate);
     }
@@ -134,12 +78,16 @@ public class CandidateSrvImpl implements CandidateSrv {
     @Override
     public CandidateDTO getById(UUID id){
         Optional<Candidate> candidate = candidateRepository.findById(id);
-        return candidate.map(mapper::candidateToCandidateDTO).orElse(null);
+        return candidate.map(mapper::candidateToCandidateDTO).orElseThrow(()-> new EntityNotFoundException("Candidate id not found : "+ id));
     }
 
 
     @Override
     public boolean delete(UUID id){
+        Optional<Candidate> candidate = candidateRepository.findById(id);
+        if (candidate.isEmpty()){
+            throw new RuntimeException("Candidate id not found : " + id);
+        }
         candidateRepository.deleteById(id);
         return true;
     }
